# Improve substr() performance by avoiding using owned string