@Dandandan 2024-12-08
I wonder if we can see perf change on TPC-H query 22 (which includes some substring expressions).