@richox 2024-12-08
> CI failure should be caused by #13686

it seems `cargo test` still fails on unrelated cases.