@Dandandan 2024-12-09
> LGTM, thanks @richox .
> 
> TPC-H query 22 uses [string_view_substr](https://github.com/apache/datafusion/blob/47569b21c50eab42771ca62fd237f429362e8a62/datafusion/functions/src/unicode/substr.rs#L201), seems not affected by this PR. cc @Dandandan

Ah makes sense